dashboard/requirements.txt
altair
pandas
streamlit
psycopg2-binary